# Conflicts:
#	src/main/java/com/warehouse/dto/response/item/ItemDetailsResponse.java
#	src/main/java/com/warehouse/mapper/ItemMapper.java
#	src/main/java/com/warehouse/repository/ItemRepository.java
#	src/main/java/com/warehouse/service/item/ItemServiceImpl.java

+++ src/main/java/com/warehouse/dto/response/item/ItemDetailsProjection.java
@@ -11,7 +11,7 @@ public record ItemDetailsProjection(
         String name,
         String category,
         int minStock,
-        int currentStock,
+        long currentStock,
         BigDecimal price,
         BigDecimal cost,
         @JsonProperty("isActive") boolean active,
